交友邀請通知功能

+++ components/friend/unaccept-friend.js
@@ -11,45 +11,60 @@ export default function UnacceptFriend({
     images,
     friendsId,
     sendFriendId,
-    deleteFriend
+    deleteFriend,
+    itemState
 }) {
     const [friendId, setFriendId] = useState('')
     const [deleteFriendId, setDeleteFriendId] = useState('')
-
+    const [itemOpen, setItemOpen] = useState(true)
 
     const handleChange = () => {
         setFriendId(friendsId)
+        setItemOpen(false)
     }
 
     const haneleDelete = () => {
         setDeleteFriendId(friendsId)
+        setItemOpen(false)
     }
 
     useEffect(() => {
+        if (friendId == '') return
         sendFriendId(friendId)
+        itemState(itemOpen)
 
     }, [friendId])
 
     useEffect(() => {
+        if (deleteFriendId == '') return
         deleteFriend(deleteFriendId)
+        itemState(itemOpen)
+
     }, [deleteFriendId])
     return (
         <>
-            <Image
-                src={`http://localhost:3002/face/${images}`}
-                className={styles.avatar}
-                alt={images}
-                width={50}
-                height={50}
-            />
-            <Btn
-                btnText='接受'
-                onClick={handleChange}
-            />
-            <Btn
-                btnText='拒絕'
-                onClick={haneleDelete}
-            />
+            {itemOpen &&
+                <div className="d-flex align-items-center justify-content-center">
+                    <Image
+                        src={`http://localhost:3002/face/${images}`}
+                        className={`${styles.avatar} me-4`}
+                        alt={images}
+                        width={50}
+                        height={50}
+                    />
+                    <Btn
+                        btnText='接受'
+                        onClick={handleChange}
+                        addClassforButton='btn-light me-3'
+                    />
+                    <Btn
+                        btnText='拒絕'
+                        onClick={haneleDelete}
+                        addClassforButton='btn-white'
+
+                    />
+                </div>
+            }
 
         </>
     )
